Move components around to improve order
