+++ frontend/src/views/bill/components/consumption/ConsumptionColors.vue
