+++ frontend/src/views/bill/components/participants/NewParticipantModal.vue
