+++ frontend/src/views/bill/components/participants/ParticipantsSelector.vue
